refactor(gulp tasks): rename gulp tasks for clarity, update gulpfile
